reorganized utils and plot, removed bad tests
- introduced a new folder called plot
- moved WeightedMarginalSamples to it's own file
- deleted old commented out tests
