+++ swyft/inference/marginalposterior.py
@@ -10,7 +10,7 @@ from swyft.inference.marginalratioestimator import MarginalRatioEstimator
 from swyft.prior import Prior, PriorTruncator, promote_to_prior_truncator
 from swyft.types import Array, MarginalIndex, MarginalToArray, ObsType
 from swyft.utils import tupleize_marginals
-from swyft.utils.weightedmarginalutils import WeightedMarginalSamples
+from swyft.weightedmarginals import WeightedMarginalSamples
 
 MarginalPosteriorType = TypeVar("MarginalPosteriorType", bound="MarginalPosterior")
