+++ swyft/inference/marginalratioestimator.py
@@ -22,7 +22,7 @@ from swyft.inference.train import get_ntrain_nvalid
 from swyft.saveable import StateDictSaveable, StateDictSaveableType
 from swyft.types import Array, Device, MarginalIndex, MarginalToArray, ObsType, PathType
 from swyft.utils.array import array_to_tensor, dict_array_to_tensor
-from swyft.utils.marginalutils import tupleize_marginals
+from swyft.utils.marginals import tupleize_marginals
 
 SchedulerType = Union[
     torch.optim.lr_scheduler._LRScheduler, torch.optim.lr_scheduler.ReduceLROnPlateau
