+++ swyft/utils/df.py
@@ -1,52 +1,28 @@
-from dataclasses import dataclass
-from typing import Optional, Tuple, Union
+from typing import Optional, Union
 
 import numpy as np
 import pandas as pd
-from toolz.dicttoolz import keyfilter
 
-from swyft.types import Array, MarginalIndex, MarginalToArray, StrictMarginalIndex
+from swyft.types import Array
 from swyft.utils import tensor_to_array
-from swyft.utils.marginalutils import tupleize_marginals
 
 
-@dataclass
-class WeightedMarginalSamples:
-    weights: MarginalToArray
-    v: np.ndarray
+def get_df_from_params(params: Array, columns: Optional[tuple] = None) -> pd.DataFrame:
+    params = tensor_to_array(params)
+    if isinstance(columns, int):
+        columns = [columns]
+    elif columns is None:
+        columns = list(range(params.shape[-1]))
+    else:
+        columns = list(columns)
+    return pd.DataFrame(params, columns=columns)
 
-    @staticmethod
-    def _select_marginal_index(marginal_index: MarginalIndex) -> Tuple[int, ...]:
-        marginal_index = tupleize_marginals(marginal_index)
-        assert (
-            len(marginal_index) == 1
-        ), "weighted marginal samples can only be recovered one index at a time"
-        return marginal_index[0]
 
-    def get_logweight(self, marginal_index: MarginalIndex):
-        marginal_index = self._select_marginal_index(marginal_index)
-        log_weight = self.weights[marginal_index]
-        return log_weight
-
-    def get_logweight_marginal(
-        self, marginal_index: MarginalIndex
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        """access the log_weight and parameter values for a marginal by index
-
-        Args:
-            marginal_index: which marginal to select. one at a time.
-
-        Returns:
-            log_weights, marginal: the log_weights and the parameter values
-        """
-        marginal_index = self._select_marginal_index(marginal_index)
-        log_weight = self.get_logweight(marginal_index)
-        marginal = self.v[:, marginal_index]
-        return log_weight, marginal
-
-    @property
-    def marginal_indices(self) -> StrictMarginalIndex:
-        return tuple(self.weights.keys())
+def get_df_dict_from_marginals(marginals: dict) -> dict:
+    dfs = {}
+    for key in marginals.keys():
+        dfs[key] = get_df_from_params(marginals[key], key)
+    return dfs
 
 
 def get_params_weights_from_weighted_marginals(
