+++ swyft/utils/marginals.py
@@ -1,11 +1,7 @@
-from typing import Optional
-
-import pandas as pd
 from toolz import keyfilter
 
 from swyft.types import Array, MarginalIndex, MarginalToArray, StrictMarginalIndex
-from swyft.utils.array import tensor_to_array
-from swyft.utils.utils import depth
+from swyft.utils.misc import depth
 
 
 def tupleize_marginals(marginals: MarginalIndex) -> StrictMarginalIndex:
@@ -51,21 +47,3 @@ def filter_marginals_by_dim(ratios: MarginalToArray, dim: int) -> MarginalToArra
         isinstance(k, tuple) for k in ratios.keys()
     ), "This function works on tuples of parameters."
     return keyfilter(lambda x: get_marginal_dim_by_key(x) == dim, ratios)
-
-
-def get_df_from_params(params: Array, columns: Optional[tuple] = None) -> pd.DataFrame:
-    params = tensor_to_array(params)
-    if isinstance(columns, int):
-        columns = [columns]
-    elif columns is None:
-        columns = list(range(params.shape[-1]))
-    else:
-        columns = list(columns)
-    return pd.DataFrame(params, columns=columns)
-
-
-def get_df_dict_from_marginals(marginals: dict) -> dict:
-    dfs = {}
-    for key in marginals.keys():
-        dfs[key] = get_df_from_params(marginals[key], key)
-    return dfs
