+++ swyft/utils/misc.py
